pyproject.toml
[project]
name = "llm-as-a-merge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.6.2",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.24",
    "termcolor>=3.3.0",
]
